+++ package.json
@@ -22,12 +22,15 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cva": "^0.0.0",
+    "dayjs": "^1.11.13",
+    "debounce": "^2.2.0",
     "i18next": "^23.16.8",
     "i18next-browser-languagedetector": "^8.0.0",
     "i18next-http-backend": "^3.0.1",
     "jotai": "^2.10.3",
     "lucide-react": "^0.460.0",
     "next-themes": "^0.4.3",
+    "qs": "^6.13.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.54.1",
@@ -41,6 +44,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.13.0",
+    "@types/qs": "^6.9.17",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react": "^4.3.3",
